refactor(admin): consolidate redundant query logic
